Reorganised files, added css on App - Please feel free to change App styles, just added it as placeholder - This includes the updates Mirim added
