+++ src/components/Button.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './Button.css';
+import '../styles/Button.css';
 
 const Button = ({ type, className, text, value, onClick }) => {
   const handleClick = () => {
