+++ src/styles/App.css
@@ -1,0 +1,9 @@
+.App {
+    width: 50%;
+    margin: auto;
+    height: fit-content;
+    border: 2px solid #2e2e2e;
+    border-radius: 2rem;
+    padding: 2rem;
+}
+
